requirements.txt
litellm>=1.40.0
python-dotenv>=1.0.0
